refactor: group pagination inputs and outputs in types/pagination
