+++ packages/platform/types/index.ts
@@ -13,4 +13,5 @@ export * from "./embed";
 export * from "./oauth-clients";
 export * from "./routing-forms";
 export * from "./me";
-export * from "./pagination";
+export * from "./pagination/pagination.input";
+export * from "./pagination/pagination.output";
