+++ packages/platform/types/pagination/pagination.output.ts
